+++ tests/test_template.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 import pytest
 from dunamai import Version
 
-from uv_dynamic_versioning import jinja, schemas
+from uv_dynamic_versioning import schemas, template
 
 
 @pytest.fixture
@@ -32,7 +32,7 @@ def test_when_rendering_basic_version_then_returns_base_version(
     version: Version,
     config: schemas.UvDynamicVersioning,
 ):
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "1.0.0"
 
@@ -46,7 +46,7 @@ def test_when_bumping_version_then_returns_bumped_version(
     config.bump = True
     config.format_jinja = "{{- base }}+r{{- revision }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "1.0.0+r2"
 
@@ -57,7 +57,7 @@ def test_when_rendering_version_with_stage_and_revision_then_returns_formatted_v
 ):
     config.format_jinja = "{{- base }}{{- stage }}{{- revision }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "1.0.0alpha1"
 
@@ -68,7 +68,7 @@ def test_when_rendering_version_with_commit_and_branch_then_returns_formatted_ve
 ):
     config.format_jinja = "{{- commit }}-{{- branch }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "message-main"
 
@@ -78,7 +78,7 @@ def test_when_rendering_version_with_timestamp_then_returns_formatted_timestamp(
 ):
     config.format_jinja = "{{- timestamp }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "20250401120000"
 
@@ -89,7 +89,7 @@ def test_when_rendering_version_with_individual_parts_then_returns_formatted_ver
 ):
     config.format_jinja = "{{- major }}.{{- minor }}.{{- patch }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "1.0.0"
 
@@ -101,7 +101,7 @@ def test_when_rendering_version_with_escaped_branch_then_returns_escaped_branch(
     version.branch = "feature/new-branch"
     config.format_jinja = "{{- branch_escaped }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "featurenewbranch"
 
@@ -112,7 +112,7 @@ def test_when_rendering_version_with_dirty_flag_then_returns_dirty_status(
     version.dirty = True
     config.format_jinja = "{{- 'dirty' if dirty else 'clean' }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "dirty"
 
@@ -125,7 +125,7 @@ def test_when_rendering_version_with_environment_variables_then_returns_env_valu
     monkeypatch.setenv("TEST_VAR", "test_value")
     config.format_jinja = "{{ env.TEST_VAR }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "test_value"
 
@@ -136,7 +136,7 @@ def test_when_rendering_version_with_serialization_functions_then_returns_serial
 ):
     config.format_jinja = "{{ serialize_pep440(base, stage, revision) }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "1.0.0a1"
 
@@ -147,7 +147,7 @@ def test_when_rendering_version_with_serialization_functions_and_bump_then_retur
 ):
     config.format_jinja = "{{ serialize_pep440(bump_version(base), stage, revision) }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "1.0.1a1"
 
@@ -159,6 +159,6 @@ def test_when_rendering_version_with_tagged_metadata_then_returns_metadata(
     version.tagged_metadata = "build123"
     config.format_jinja = "{{ tagged_metadata }}"
 
-    result = jinja.render_jinja(version, config)
+    result = template.render_jinja(version, config)
 
     assert result == "build123"
